separate service layer to routes
