+++ backend/src/services/export.service.ts
@@ -1,11 +1,10 @@
-import { FastifyRequest } from "fastify";
-import { ExportInput, ExportResult } from "@/types/export.types";
+import { FastifyReply, FastifyRequest } from "fastify";
+import { ExportFormat, ExportInput, ExportResult, HotspotData } from "@/types/export.types";
 import { AuthService, getAuthService } from "./auth.service";
 import { getNotificationHubService } from "./notificationHub.service";
 import { GoogleDriveStorageProvider } from "@/providers/storage/GoogleDriveStorageProvider";
 import { ExportFormatterFactory } from "@/providers/export/ExportFormatterFactory";
 import { google } from "googleapis";
-import { auth } from "firebase-admin";
 import { AuthProviderFactory } from "@/providers/auth/AuthProviderFactory";
 import "@/config/env"; // Ensure environment variables are loaded
 
@@ -26,12 +25,62 @@ export class ExportService {
     this.notificationHub = notificationHub;
   }
 
+  async buildExportOptions (request: FastifyRequest, reply: FastifyReply):Promise<ExportInput>{
+    const formFields: Record<string, string> = {};
+    let fileBuffer: Buffer | null = null;
+    let filename = '';
+    let mimetype = '';
+
+    // Iterate through ALL multipart parts
+    const parts = request.parts();
+    for await (const part of parts) {
+      if (part.type === 'file') {
+        // Handle file part
+        fileBuffer = await part.toBuffer();
+        filename = part.filename;
+        mimetype = part.mimetype;
+      } else if (part.type === 'field') {
+        // Handle text field parts
+        formFields[part.fieldname] = part.value as string;
+      }
+    }
+
+    if (!fileBuffer || fileBuffer.length === 0) {
+      return reply.status(400).send({ error: 'Image or picto file required' });
+    }
+
+    const format = formFields.format as ExportFormat || 'picto';
+    let annotations: HotspotData[] | undefined = undefined;
+
+    if (format === "raw" && formFields.annotations) {
+      try {
+        annotations = JSON.parse(formFields.annotations);
+      } catch (err) {
+        return reply.status(400).send('Error: Invalid annotations JSON');
+      }
+    }
+    const options = { 
+      format,
+      fileName: formFields.fileName || undefined,
+      folderName: formFields.folderName || undefined,
+      includeMetadata: formFields.includeMetadata === 'true'
+    }
+
+    return {
+      fileBuffer,
+      annotations,
+      options,
+    }
+  }
+
   /**
    * Export file to Google Drive
    * Handles auth, storage provider setup, and format selection
    */
-  async exportToGoogleDrive(request: FastifyRequest, input: ExportInput): Promise<ExportResult> {
+  async exportToGoogleDrive(request: FastifyRequest, reply: FastifyReply){
     try {
+      const export_input :ExportInput= await this.buildExportOptions(request,reply);
+
       // Ensure valid auth and get access token
       const accessToken = await this.authService.ensureConnection(request);
       const scope = this.authService.getSessionScope(request);
@@ -51,7 +100,7 @@ export class ExportService {
       const storage = new GoogleDriveStorageProvider(oauth2Client);
 
       // Create export folder
-      const folderName = input.options.folderName || "360° Image Annotations";
+      const folderName = export_input.options.folderName || "360° Image Annotations";
       const folderId = await storage.createFolder(folderName);
 
       // Notify client: folder created
@@ -62,12 +111,12 @@ export class ExportService {
       });
 
       // Get formatter and export
-      const format = input.options.format;
+      const format = export_input.options.format;
       const formatter = ExportFormatterFactory.getFormatter(format);
       const exportResult = await formatter.export(
-        input.fileBuffer,
-        input.annotations,
-        input.options,
+        export_input.fileBuffer,
+        export_input.annotations,
+        export_input.options,
         storage,
         folderId,
         (event, data) => {
@@ -95,10 +144,9 @@ export class ExportService {
         },
       );
 
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : "Export failed",
-      };
+      reply.status(500).send({
+        error: error instanceof Error ? error.message : 'Export failed'
+      });
     }
   }
 }
